+++ packages/search-pro/rollup.config.ts
@@ -15,10 +15,18 @@ export default [
       copy: [["client/styles", "client"]],
     }
   ),
-  ...bundle("worker/index", {
+  ...bundle("worker/minify", {
     resolve: true,
     dts: false,
     external: [/^@internal\//],
+    replace: {
+      // eslint-disable-next-line @typescript-eslint/naming-convention
+      __VUEPRESS_SSR__: false,
+    },
+  }),
+  ...bundle("worker/original", {
+    resolve: true,
+    dts: false,
     replace: {
       // eslint-disable-next-line @typescript-eslint/naming-convention
       __VUEPRESS_SSR__: false,