+++ packages/search-pro/src/worker/original.ts
@@ -1,0 +1,13 @@
+import { getResults } from "../client/utils/index.js";
+import { type SearchIndex } from "../shared/index.js";
+
+const searchIndex: SearchIndex = <SearchIndex>JSON.parse(
+  // @ts-ignore
+  SEARCH_PRO_INDEX as string
+);
+
+self.onmessage = ({
+  data,
+}: MessageEvent<{ query: string; routeLocale: string }>): void => {
+  self.postMessage(getResults(data.query, searchIndex[data.routeLocale]));
+};