+++ packages/theme/src/node/plugins/blog/category.ts
@@ -1,4 +1,5 @@
 import { type GitData } from "@vuepress/plugin-git";
+import { isArray } from "@vuepress/shared";
 import { type BlogCategoryOptions } from "vuepress-plugin-blog2";
 
 import { defaultPageSorter } from "./utils.js";
@@ -27,7 +28,11 @@ export const getBlogCategoryCategory = (
     >
   >{
     key: "category",
-    getter: ({ routeMeta }) => routeMeta[ArticleInfoType.category] || [],
+    getter: ({ routeMeta }) => {
+      const category = routeMeta[ArticleInfoType.category];
+
+      return isArray(category) ? category : category ? [category] : [];
+    },
     sorter: defaultPageSorter,
     path: options.category,
     layout: "BlogCategory",
@@ -64,7 +69,11 @@ export const getBlogTagCategory = (
     >
   >{
     key: "tag",
-    getter: ({ routeMeta }) => routeMeta[ArticleInfoType.tag] || [],
+    getter: ({ routeMeta }) => {
+      const tag = routeMeta[ArticleInfoType.tag];
+
+      return isArray(tag) ? tag : tag ? [tag] : [];
+    },
     sorter: defaultPageSorter,
     path: options.tag,
     layout: "BlogCategory",